moved `apex.py` and `separating_set.py` to `assorted`
