+++ pyrigi/graph/assorted/separating_set.py
