+++ pyrigi/graph/graph.py
@@ -21,7 +21,8 @@ from pyrigi.misc.misc import _generate_category_tables
 from pyrigi.plot_style import PlotStyle
 
 from . import _input_check as _graph_input_check
-from . import apex, general, separating_set
+from . import general
+from .assorted import apex, separating_set
 from .constructions import extensions
 from .rigidity import generic as generic_rigidity
 from .rigidity import global_ as global_rigidity
