+++ test/graph/assorted/test_apex.py
@@ -4,7 +4,7 @@ from itertools import product
 import networkx as nx
 import pytest
 
-import pyrigi.graph.apex as apex
+import pyrigi.graph.assorted.apex as apex
 import pyrigi.graphDB as graphs
 from pyrigi.graph import Graph
 from test.graph.test_graph import TEST_WRAPPED_FUNCTIONS
